refactor: req/res 디렉토리 구분
