+++ src/main/java/sleppynavigators/studyupbackend/presentation/notification/dto/request/FcmTokenRequest.java
@@ -1,4 +1,4 @@
-package sleppynavigators.studyupbackend.presentation.notification.dto;
+package sleppynavigators.studyupbackend.presentation.notification.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
